+++ lib/Braintree/Configuration.php
@@ -52,8 +52,26 @@ class Configuration
             if ($kind == 'privateKey') {
                 $this->_privateKey = $value;
             }
+            if ($kind == 'proxyHost') {
+                $this->_proxyHost = $value;
+            }
+            if ($kind == 'proxyPort') {
+                $this->_proxyPort = $value;
+            }
+            if ($kind == 'proxyType') {
+                $this->_proxyType = $value;
+            }
+            if ($kind == 'proxyUser') {
+                $this->_proxyUser = $value;
+            }
+            if ($kind == 'proxyPassword') {
+                $this->_proxyPassword = $value;
+            }
             if ($kind == 'timeout') {
                 $this->_timeout = $value;
+            }
+            if ($kind == 'sslVersion') {
+                $this->_sslVersion = $value;
             }
             if ($kind == 'acceptGzipEncoding') {
                 $this->_acceptGzipEncoding = $value;
@@ -195,12 +213,15 @@ class Configuration
     }
 
     /**
+     * @deprecated Use isUsingInstanceProxy instead.
      * Specifies whether or not a proxy is properly configured
      *
      * @return bool true if a proxy is configured properly, false if not
      */
     public static function isUsingProxy()
     {
+        // NEXT_MAJOR_VERSION Remove this method and rename isUsingInstanceProxy to isUsingProxy
+        trigger_error('DEPRECATED: Use isUsingInstanceProxy instead.', E_USER_DEPRECATED);
         $proxyHost = self::$global->getProxyHost();
         $proxyPort = self::$global->getProxyPort();
         return !empty($proxyHost) && !empty($proxyPort);
@@ -223,6 +244,7 @@ class Configuration
     }
 
     /**
+     * @deprecated Use isAuthenticatedInstanceProxy instead.
      * Specified whether or not a username and password have been provided for
      * use with an authenticated proxy
      *
@@ -230,6 +252,8 @@ class Configuration
      */
     public static function isAuthenticatedProxy()
     {
+        // NEXT_MAJOR_VERSION Remove this method and rename isAuthenticatedInstanceProxy to isAuthenticatedProxy
+        trigger_error('DEPRECATED: Use isAuthenticatedInstanceProxy instead.', E_USER_DEPRECATED);
         $proxyUser = self::$global->getProxyUser();
         $proxyPwd = self::$global->getProxyPassword();
         return !empty($proxyUser) && !empty($proxyPwd);
@@ -416,7 +440,7 @@ class Configuration
         $this->_sslVersion = $value;
     }
 
-    private function getSslVersion()
+    public function getSslVersion()
     {
         return $this->_sslVersion;
     }
@@ -530,6 +554,20 @@ class Configuration
             return 443;
         }
         return getenv("GRAPHQL_PORT") ?: 8080;
+    }
+
+    public function isUsingInstanceProxy()
+    {
+        $proxyHost = $this->getProxyHost();
+        $proxyPort = $this->getProxyPort();
+        return !empty($proxyHost) && !empty($proxyPort);
+    }
+
+    public function isAuthenticatedInstanceProxy()
+    {
+        $proxyUser = $this->getProxyUser();
+        $proxyPwd = $this->getProxyPassword();
+        return !empty($proxyUser) && !empty($proxyPwd);
     }
 
     /**